examples: Added in 3 unstructured case examples.
This update adds in three unstructured cases. These
are straight ports of the followingstructured simulations,
1. cyl50
2. hakkinen-swbli
3. laminar-flat-plate
Note that currently the no_slip_fixed_T boundary condition
requires access to kOmegaWall (but doesn't actually use it) which is not
currently implemented for the unstructures solver. To get around this
the usg cases define a temporary no_slip_fixed_T boundary condition
in the user-script.

Also note that for the hakkinen case when using the current fixed_PT
outflow boundary condition on the unstructured solver causes an anomaly
along the first column of cells. It is reccomended that this be changed to a simple
extrapolate out boundary condition.

Finally, the cyl50-usg case includes a tcl reference test to check the unstructured
viscous code when changes are made (this is not added to the suite of tests due to the
lengthy run time ~30mins on a typical workstation).