eilmer-gpu: modification to prep-gpu-chem-kernel.
Added in ability to handle a loss/production rate equation
when a species is not used in a reaction mechanism (i.e.
the case when there are columns filled with all 0's
in the stoichiometric matrix). For such cases the
prep file should fill these rate equations with a 0 in
the gpu-kernel.

The Stanford model is an example of such a reaction mechanism.