Eilmer: remove the arbitrary length limit on point clouds for least-squares.
The price to pay for making the workspace arrays dynamic is that some of the
functions can no longer be guaranteed not to allocate memory.  Indeed, the
set up of the workspaces will allocate memory on first use of each workspace
and, in the case of the common workspace, every time a larger point cloud is
encountered.  Eventually, no more allocations will happen because that largest
cloud will be encountered, some time on the first pass through all points.