Eilmer: Eliminate the hierarchy of cell types and return to having only FVCell.
The BasicCell class didn't seem to offer much benefit for the confusion it
caused.  I'm also hoping that reducing the classes to just the FVCell class
will allow the D compiler to generate less/faster code since the type of all
cells in all parts of the code is now known at compile time.