Eilmer: adjust the tolerances of the sod shock tube problem, unstructured grid.
Increase the tolerance just a little in the expanded-gas region so that the
test does not fail on small changes as we fiddle with reconstruction schemes
for the convective fluxes.  The solution, with Kyle's new reconstruction scheme,
looks fine.  Large changes are unexpected so we want the test to indicate those.