Eilmer: experiment with expanding the Vector3 gradient values as scalar components.
This commit is to record the code and result of that experiment, before going
back to the earlier version.

The idea was to see if dealing with the scalar components directly was faster
than using the Vector3 objects.  The cone20-usg test seems to run a little
faster at 37 seconds, down from 42 seconds, compiled for debug on my AMD
workstation (helmholtz).  So there is a speed benefit.

The cost is that the code hurts to read.  The functions used in the mixins have
doubled and now look much like the much feared m4 macros that we used to suffer
in Eilmer3.