eilmer: Updates to the Barth Limiter.
This push adds in some algorithmic updates to the Barth limiter.
We try to do as much work internally without relying on borrowed methods.
Doing this has saved us a large chunk of time, and has brought the
Barth limiter to almost the same cost as the van albada limiter.

Note this push makes the Barth limiter the default limiter for now whilst
I am working heavily on the limiter implementations.