Eilmer: remove the experiment of computing flow gradients from a pair of cells.
This commit is the first of a small set of commits to clean up the gradient
calculation module.

Will try to bring this approach back as a simple direct viscous flux calculator
rather than trying to fit within the framework of computing all gradients and
then calculating viscous fluxes from those gradients.