Eilmer: finish refactoring the least-squares flow gradients calculation.
It is now set up as arrays of weights precomputed using geometric information.
The flow gradients are now computed as weighted sums.

We've also eliminated more of the memory allocations within the functions,
but at the cost of having a fixed upper limit on the numbers of points
in the clouds.