Eilmer examples: added the Sawada sphere, case 3.
This is a direct port of the Eilmer3 example and makes use of the uniform
look-up table code ported by James Burgess some months ago.  Thanks James!

Shock stand-off, as extracted by the awk program using the stagnation-line data,
is 2.64mm.  This is close enough for me.

Finally, note that the test script expects to find the look-up table file,
as prepared by James, in $HOME/dgd/src/gas/sample-data/