gas: Removed default values in private area of IdealGas class.
In last week's rearrangement of the IdealGas class, I stripped
out the default constructor which would set up ideal air if
no other information was given. Since the constructor was gone,
the default values for air in the private variables were a bit
of a red herring. These have been removed.