Gas: copy the bracketing and solver functions into therm_perfect_gasmix_eos
...so that we can use simple function calls rather than construction and
passing delegates when the temperature is updated from the energy.
This temperature update is done often and from separate threads when
the code is run on multiple CPUs.  We believe that the construction of delegates
requires memory allocation and will hinder the running of the threads.

Even though we really like the neat solution offered by the delegate, we want
the parallel speed of the code even more.  Now that the change has been made,
we need to go measure to be sure that we haven't ended up with ugly and slow code.