Geom: added render function for parametric volumes.
Allowed the surface renderer to draw a single large polygon (which works
well in 2D and is compact) or renders the surface as facets (which might work
better for curved surfaces in 3D, at greater cost).

And, reduced the resolution of the discretization from 100 to 30 so that the
number of facets is not so large that the SVG files become too big to handle
in inkscape.