Eilmer: doing the recovery of bad cells carefully is hard work.
Update the correct flow-level index for the conserved quantities because we are
checking at the end of each stage of the gasdynamic update.

Also, rescale the averaged mass-fraction values because, with different density
values in the neighbouring cells, they probably won't correctly sum to 1.0
after taking a simple average of neighbouring mass-fraction values.