eilmer: Introduce a formal scaling matrix for the GMRES solver.
This is just a cosmetic change to the code. The scaling was already
being performed but I've just reworked the somewhat ad hoc scaling
into an array arrangement. The code now lines more nicely with
the notation in my notes and in journal papers.

There should be no change in functionality in this commit.