eilmer/examples: Changed number of iterations to 60 for sharp-wedge case.
It tutns out you can really drive the residuals down quite quickly with the
Newton updates if you're prepared to do some more work at each step
on solving the matrix system to give the updates. The number of inner
iterations on the solver was increased from 30 to 60.

The plot script was updated so that relative residuals
are properly plotted.