eilmer: Implemented larger than nearest-neighbour point clouds.
The nearest-neighbour approach to storing point clouds for use with the
least-squares gradient method has shown to not be adequate for use
on unstructured triangular meshes generated using a Delaunay algorithm.
It was found that in some instances the point clouds did not provide
a good enough distribution in all dimensions, and this lead to
bad cells in the MMS test cases.

To make the unstructured solver more robust, we now check the standard
deviation of the cell clouds prior to assembling and inverting the least-
squares matrix. If the standard deviation in any of the dimensions is not
larger than half the cells average face length than more cells are added
to the cloud. The cells which are appended are the face-neighbours of the
initial nearest neighbour point cloud cells.