eilmer: Refactored usg spatial derivative cloud construction.
The code for constructing the spatial derivative clouds for the
unstructured solver was getting needlessly complex. In this revision
the code is completely refactored and restructured. Now every internal
faces spatial derivative cloud consists of the interface and all the
faces of the neighbouring two cells plus the neighbouring cell
centres.