eilmer/sharp-wedge-6-degrees: Update parameter names in input script.
With the introduction of a preconditioner, the terminology for referring
to iterations was becoming murky. I opted to make this (somewhat arbitrary)
distinction:

The Newton updates, which are at the outer most level, are called 'steps'.

The Newton updates require the solution of a linear system. That linear
system is solved with an iterative method. The steps within that iterative
method are called 'iterations'. In particular, the user can control the
maximum number of outer iterations.

The linear solver uses a preconditioner on each iteration. The preconditioner
itself is some small number of iterations solving the unpreconditioned
system using a GMRES method. This small number of iterations are referred
to as the inner iterations. An iterative preconditioner for an iterative
matrix solver has been called 'sqared preconditioning' by some because
of the iterations within iterations.

To reflect this naming convention, the parameters in the wedge.lua
script have been updated.

I also added a gnuplot plot that will pop-up at run time and
display the progression of the residuals. The graph will disappear
when the user presses "x".