eilmer: Replaced "linear_reduction" tolerance reduction strategy with "geometric"
The tolerance reduction strategy (etaStrategy in code) is a means to control
the linear solver tolerance during simulation. The idea is to tighten the
tolerance on the linear solver as steady-state convergence is approached.
The "linear_reduction" strategy didn't really make a whole lot of sense once
I gave it some more thought. We are hoping for quadratic convergence or
at least super-linear for this inexact Newton method and the linear reduction
in tolerance doesn't help much to that end.

Instead, a geometric progression is now available as tolerance reduction
strategy. The tolerance setting, eta, will given by the geometric progression:

eta = eta0, eta0*r, eta*r^2, eta0*r^3, ....

where 0 > eta0, r > 1.

In the code, 'r' is set as "eta_ratio_per_step".