Geom+Eilmer: find_enclosing_cell now uses an array of bins of cells.
The intention is to reduce the time taken for most searches by reducing the
set of candidate cells to those near the given point.  The cells are sorted
into the array of bins just after the grid is read in.

This all happens within the Grid class and behind the scenes, so we shouldn't
have to do anything/much to gain from it in other parts of the code.