grid_utils: Improved efficiency for ugrid_partition.
The first version of the partitioner applied the naive approach to
assign each domain boundary face to the correct partition this was done
by looping over each boundary face and then searching the entire grid
to determine what partition it belongs to. This update adds a list to
each face object in which we store the partitions the face belongs
to at the time it is created.

The partitioning of large meshes is now tractable.

NB: it has been noticed that some partitioned meshes cause Eilmer4 to
crash at the point of reading in the vertices from the su2 file.
This is due to a bug in transforming the global node id to the local
node id. This is currently being fixed.