Eilmer: Shared-memory parallel Newton-Krylov update... STAGE 3 (final).
In this stage, I have added in the parallel foreach to
handle block updates in parallel where appropriate.
This seems to work for the simple Mach 10 wedge flow.
I think there are more efficiencies to be gained in
the implementation, but I'm pleased this first-pass
effort has turned out well.