Gasdyn: cleaned up a little notational difference between Momar and me.
Momar had changed the u2_u1_obl function to return the ratio of normal
velocities, but the Python code for the conical-flow calculator assumed that
this name still referred to the speed ratio across the shock.

We now use the name Vn2_Vn2_obl for the ratio of normal velocities and
V2_V1_obl for the speed ratio.