bc: Moved setting the ghost cell position to the mapped cell position to after we write out the cell infomation.
We have a verbose setting for the mapped cells method where the original ghost cell and mapped cells position
are written out to the screen for debugging. The update in r796 caused the ghost cell position to be updated
before this step, which isn't helpful for debugguing. This update moves the line of code that sets the ghost
cell position to that of the mapped cells until after the debugging print statements.