eilmer: Added mapped cells from file
This update contains all the modifications to allow ghost cells to be
filled with mapped cells from a file generated by the ugrid_partition
program.

Cleaning up of the new code will take place soon and be pushed as a separate
revision. This update acts as a known good state until then.