Eilmer/Geom: Add functions to check volumes of structured grids at prep stage.
There is a new low-level method in the Grid class:
double cell_volume(size_t idx);
This will return the cell volume so that we can check
that it is positive. This has been completed for
structured grids.

TODO: implement cell_volume() method for unstructured grids.

There is a high-level function that a user might insert into
his/her prep script in order to force a check of the cell
volumes at preparation stage. This function must be called
AFTER block construction since it uses the blocks to get access
to the grids. To use this in the prep script:

cellsOK, badCells = checkCellVolumes{badCellLimit=20}

The function returns a boolean and a list of bad cells.
The list is limited to the number set by badCellLimit
(which defaults to 20 if not set). If the badCellLimit is
exceeded while doing the check, the function just returns
with the identities of bad cells up to that point. The list of
badCells contains table entries of the form {blk-indx, cell-indx}.