Gasdyn: some small improvements to oblique-shock relations.
(1) Go back to the Python expression for speed-ratio because it doesn't need
both beta and theta as input.  That way we don't have to be sure that we pass
consistent values into the function.

(2) In most cases, the exception for subsonic inflow should test the normal
component of the Mach number.

(3) Made the capture of the sign of the incoming angle a little tidier (but
tidier only in the eye of this code-writer, probably).