eilmer: Fixed FlowSolution bug added in with r883.
In revision 883, the FlowSolution constructor was edited to
allow for the simulation config file to be read. The original
code relied on using the globalconfig name which was defaulting
to "job" when running custom-post, hence breaking a good number
of the automated tests.

This commit changes the filename to use the jobName, which is
available to the flowsolution constructor.