Eilmer: partly implemented InflowBC_StaticProfile
and then realized that I was about to make yet another function to scan
the cell data from a string.  We will return to this new boundary condition
when that redundant code for writing to and scanning from string is
cleaned up.  The partly implemented BC should not used for the moment.