Eilmer: connect code for new boundary condition InFlowBC_StaticProfile
The new boundary condition can be applied and it mostly works, however,
it makes use of only the first point in the profile and thus looks like
the usual Supersonic inflow boundary condition.  The cell matching code
needs to be completed so that a number of points can be used.

The format of the profile file is just the current output from the post-
processing extract-line option.  This make it easy to set up a profile.

Also, added an example of the sharp cone restarted with the new boundary
condition.  The runB.sh script sets up the profile file from a previous
cone20 run.