Eilmer: adjust the velocity for InFlowBC_StaticProfile.
Also, change the naming of the options to:
"xyz-to-xyz" closest distance considering all components of position.
"xyA-to-xyA" A means any-z.
"AyA-to-AyA" any-x and any-z, closest y.
"xy-to-xR" R is in the yz plane.
"Ay-to-AR" any-x and R is in the yz plane.
For the last two options, the velocity will be adjusted.