kinetics: New form of pressure-dependent rate constant implemented: Yungster-Rabinowitz.
Cor Lerink is working on the methane chemistry test case in a paper by
Yungster and Rabinowitz (1994). In that paper, they have a form of
pressure-dependent rate constant that differs slightly in form from those
that are supported by Chemkin. It seemed cleanest to implement this form
as a separate rate constant rather than try to shoehorn the rate to fit
into one of the existing forms. An example of the new rate constant
specification is below. This corresponds to Reaction 33 in Yungster and
Rabinowitz's paper.

Reaction{
   'CH4 (+ M) <=> CH3 + H (+ M)',
   fr={'pressure dependent',
        kInf={A=7.05e16, n=-0.558, C=52788.95},
        k0={A=1.19e35, n=-4.911, C=53829.68},
        YR={a=0.555, b=405.62, c=4580.87}
   },
   label='r33'
}