Gas: more progress on CEAGas.
We can now write the tmp.inp file for "pT" updates and can run the CEA2 program.
Next to do is the scanning of the tmp.out file.  We will need to think about
storing some data in the GasState class so that we can preserve the "const"ness
of the gas model methods.