gas: When using lua-wrapped speciesName() method indices start from 0.
I had an inconsistency in the species indices. The speciesName() method
accepts an index and returns a species name. I was expecting 1-offset indices and
converting them to 0-offset internally. However, all other user-facing
species indices are 0-offset. I think I had tried to make the Lua-wrapped
interface more Lua-like with 1-offset for indices, but I was inconsistent in
doing that.

In general, this problem has been solved by using named species in the
table interface. There are few functions that require the species index
rather than a name. Where the species index is required, it will use the
same 0-offset as the internal D code.

[TODO] I should think about converting to using species names in
all user-facing code.