gas: User can now select between CEA or GRIMECH coefficients for thermo.
When using the thermally perfect gas model, the default behaviour
will be to look for the CEA coefficients. However, there is an
option where the user can specify that they would prefer to use
the GRIMECH coefficients (which are basically the *old* version
of the JANNAF coefficients). I expect this option would be selected
when someone wanted to do a perfect replication of a GRIMECH
reaction mechanism.

To select GRIMECH coefficients, the user adds that preference in
the 'options' table as part of a prep-gas input file:

model = 'thermally perfect gas'
species = {'CH4', 'CH3', 'CH2', ... long list of hydrocarbons ..., }
options = {database='prefer-grimech'}

If the GRIMECH coefficients are NOT available, then prep-gas
will issue a warning and default back to the CEA coefficients.