eilmer: Added in compute-loads-on-boundary funcationality.
This is the first addition of the compute-loads-on-boundary function.

This work similar to the history function, the user sets a dt_loads in the
user input script, the solver then writes out the loads on specified domain
boundaries in the loads directory.

The user currently specifies whether they want a domain boundary to have
loads computed and stored by assigning the boundary a group name. If a
boundary does not have a group assigned, then it will not compute nor
store the loads. The current file naming structure is: solver-tindx_group-name.dat.